+++ lib/ModelMixins/UrlMixin.ts
@@ -6,7 +6,7 @@ import UrlTraits from "../Traits/UrlTraits";
 
 type UrlModel = Model<UrlTraits>;
 
-export default function UrlMixin<T extends Constructor<UrlModel>>(Base: T) {
+function UrlMixin<T extends Constructor<UrlModel>>(Base: T) {
   class UrlMixin extends Base {
     @computed
     get uri(): uri.URI | undefined {
@@ -19,3 +19,12 @@ export default function UrlMixin<T extends Constructor<UrlModel>>(Base: T) {
 
   return UrlMixin;
 }
+
+namespace UrlMixin {
+  export interface UrlMixin extends InstanceType<ReturnType<typeof UrlMixin>> {}
+  export function isMixedInto(model: any): model is UrlMixin {
+    return model && model.isGroup;
+  }
+}
+
+export default UrlMixin;