+++ lib/Models/WebProcessingServiceCatalogFunction.ts
@@ -135,7 +135,7 @@ export default class WebProcessingServiceCatalogFunction extends CatalogMemberMi
       Identifier: this.identifier
     });
 
-    return proxyCatalogItemUrl(this, uri.toString(), this.cacheDuration);
+    return proxyCatalogItemUrl(this, uri.toString());
   }
 
   /**
@@ -151,7 +151,7 @@ export default class WebProcessingServiceCatalogFunction extends CatalogMemberMi
       request: "Execute",
       version: "1.0.0"
     });
-    return proxyCatalogItemUrl(this, uri.toString(), this.cacheDuration);
+    return proxyCatalogItemUrl(this, uri.toString());
   }
 
   async forceLoadMetadata() {